Port latest private Stadium 1 decomp findings

+++ src/gb_audio_render.h
@@ -6,6 +6,6 @@
 void GbAudio_GenerateSamples(s16*, s32, u32, f32);
 void GbAudio_QueueRegisterWrite(u16, u8, u16);
 u8 GbAudio_GetRegisterValue(u16);
-void func_800498A8(u32, u8, u32);
+void func_800498A8(s32, s32, s32);
 
 #endif // _4A3E0_H_
